# Conflicts:
#	env_var
#	simplyblock_core/models/nvme_device.py
#	simplyblock_core/storage_node_ops.py

+++ simplyblock_web/utils.py
@@ -79,7 +79,7 @@ def parse_size(size_string: str):
             else:
                 print(f"Error parsing size: {size_string}")
                 return -1
-            return size_number * math.pow(one_k, multi)
+            return int(size_number * math.pow(one_k, multi))
         else:
             return -1
     except:
